pyproject.toml
[project]
name = "hermitclaw"
version = "0.1.0"
description = "A tiny AI creature that lives in a folder on your computer"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "websockets>=12.0",
    "openai>=1.50.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
]
